+++ packages/editor/editor-core/src/__tests__/integration/table/resize.ts
@@ -28,7 +28,7 @@ import { TableCssClassName } from '../../../plugins/table/types';
 BrowserTestCase(
   'Can resize normally with a rowspan in the non last column.',
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -42,14 +42,14 @@ BrowserTestCase(
     await resizeColumn(page, { cellHandlePos: 2, resizeWidth: 50 });
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
 BrowserTestCase(
   'Can resize normally with a rowspan and colspan',
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -63,14 +63,14 @@ BrowserTestCase(
     await resizeColumn(page, { cellHandlePos: 22, resizeWidth: -50 });
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
 BrowserTestCase(
   'Can resize normally on a full width table with content',
   { skip: ['ie', 'edge', 'firefox', 'safari'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -84,7 +84,7 @@ BrowserTestCase(
     await resizeColumn(page, { cellHandlePos: 2, resizeWidth: -100 });
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
@@ -93,7 +93,7 @@ BrowserTestCase(
   // Firefox clicks the wrong cell with the TOP_LEFT_CELL
   // Needs some digging.
   { skip: ['ie', 'firefox'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -109,14 +109,14 @@ BrowserTestCase(
     await insertBlockMenuItem(page, 'Block macro (EH)');
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
 BrowserTestCase(
   `Created column should be set to ${tableNewColumnMinWidth}px`,
   { skip: ['ie', 'safari'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -130,14 +130,14 @@ BrowserTestCase(
     await insertColumn(page, 1);
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );
 
 BrowserTestCase(
   "Can't resize the last column of a table with dynamic sizing enabled.",
   { skip: ['ie'] },
-  async (client: any) => {
+  async (client: any, testName: string) => {
     const page = await goToEditorTestingExample(client);
 
     await mountEditor(page, {
@@ -152,6 +152,6 @@ BrowserTestCase(
     await resizeColumn(page, { cellHandlePos: 10, resizeWidth: -100 });
 
     const doc = await page.$eval(editable, getDocFromElement);
-    expect(doc).toMatchDocSnapshot();
+    expect(doc).toMatchCustomDocSnapshot(testName);
   },
 );