+++ packages/home/notification-log-client/src/__tests__/unit/NotificationLogClientSpec.ts
@@ -1,5 +1,5 @@
 import * as fetchMock from 'fetch-mock/src/client';
-import { version as npmPackageVersion } from '../../../package.json';
+import { version as npmPackageVersion } from '../../version.json';
 import {
   NotificationLogClient,
   NotificationCountResponse,